+++ ErsatzTV.Application/Libraries/Queries/GetConfiguredLibraries.cs
@@ -3,5 +3,5 @@ using MediatR;
 
 namespace ErsatzTV.Application.Libraries.Queries
 {
-    public record GetAllLibraries : IRequest<List<LibraryViewModel>>;
+    public record GetConfiguredLibraries : IRequest<List<LibraryViewModel>>;
 }